Generation of Numbers Based on Diff. Level Added
Generation of numbers based on difficulty level has been added. Code is
very messy however due to the trial and error attempts that were done to
accomplish this. The passing of the math type still has to be
implemented, however, and the code will have to be cleaned up and
commented at a later date.