Basic Quiz Functionality Now Completely Working!
The basic quiz functionality of checking user-inputted answers is now
completely working! The selected math type is now passed down to the
gameRun.html page, so the web app now loads equations according to the
user's math type choice.

Code still needs to be cleaned up, and a modal indicating that the user
either won (when reaching the goal points) or lost (if the number of
seconds from the timer reaches 0) still needs to be implemented.  Adding
an animated graphic to indicate whether a user answered an equation
correctly or not needs to be added as well.