Added a bunch more functionality
There is now a win/lose pop up that appears when you win or lose. It
displays the info for that game session in a table. It has a main menu
button and a submit score button (doesn't work yet tho). Also removed
time elapsed from the GUI - it's still saved internally, but it was
making the GUI ugly.