Use Sprite for Diffic. Level Checkbox
Instead of using two separate images for the difficulty level
checkboxes, a single image (or sprite) is now used. This saves an HTTP
request which is very important. The dta URI of the sprite is being used
here instead of the sprite itself to test to see if there is a small
load speed increase.