Removed Button Gray Borders & Changed TimeDispl.
The gray area surrounding the buttons has been removed. It was the
result of a CSS issue, in which the background: CSS property by itself
removes the gray border when an image URL is followed; when using the
background-image property instead, it another property,
background-color, must be defined as transparent in order to a gray area
not to appear.

The time displayed has been changed to a whole number and not a decimal
number. This removes the lag between the numbers changing as well as any
potential jumps. However, deciding the score via the time is still
determined by a decimal number of time.